Tests/LibWeb/Text/expected/css/registered-property-currentcolor.txt

Inner background color is: rgb(0, 128, 0)
